requirements.txt
Django==5.2.9
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dateutil==2.8.2
pytz==2024.1
